giving up for the day/week

+++ app.py
@@ -1,11 +1,14 @@
+import datetime
 import os
 import time
-import datetime
 
 import requests
 from dotenv import load_dotenv
 from flask import Flask, request
+from flask_caching import Cache
 from slack_sdk import WebClient
+
+print(f"{datetime.datetime.now()}: This is printed from outside any function.")
 
 # get the environment variables from .env
 load_dotenv()
@@ -16,22 +19,22 @@ client = WebClient(token=bot_token)
 headers = {"Authorization": "Bearer " + bot_token}
 
 app = Flask(__name__)
+cache = Cache(app, config={'CACHE_TYPE': 'simple'})
+
 @app.route('/slack/events', methods=['POST'])
 def slack_event_handler():
-
-    # get the time now
-    print(f"{datetime.datetime.now()}: Request received")
+    print(f"{datetime.datetime.now()}: Request received. Starting event handler.")
 
     # get the request data in json format
     request_data = request.get_json()
-
+    
     # authenticate URL
     if "challenge" in request_data:
         return request_data["challenge"]
 
     else:
         event_id = request_data["event_id"]
-        
+
         event_data = request_data["event"]
         user_id = event_data["user"]
         reaction_type = event_data["reaction"]
@@ -39,50 +42,61 @@ def slack_event_handler():
         item_data = event_data["item"]
         channel_id = item_data["channel"]
         timestamp = item_data["ts"]      
-      
+    
         print(f"{datetime.datetime.now()}: Event: {event_id}; User: {user_id}; Rxn: {reaction_type}; Chan: {channel_id}; TS: {timestamp}.")
 
-        # handle emoji reactions
-        if reaction_type != "label":
-            print(f"{datetime.datetime.now()}: Can't compute :{reaction_type}: reaction!")
-        else:  
-            conversation = client.conversations_replies(
-                channel=channel_id,
-                ts=timestamp
-            )
-            messages = conversation.get("messages")
-            root_message = messages[0]
-
-            if "files" in root_message:
-                print(f"{datetime.datetime.now()}: File found.")
-                file_id = root_message["files"][0]["id"]
-                file_vtt = get_file_info(file_id)
-                save_location = event_id + '.vtt'
-                vtt_file_for_conversion = download_vtt_file(file_vtt, save_location)
-                txt_file_output = event_id + ".txt"
-                finished_txt_file = convert_vtt_to_labels(vtt_file_for_conversion, txt_file_output)
-                client.files_upload_v2(
-                channel=channel_id,
-                thread_ts=timestamp,
-                initial_comment="Thanks for using LabelMaker! Here's your labels file:",
-                file=finished_txt_file
-            )
-                os.remove(vtt_file_for_conversion)
-                os.remove(finished_txt_file)
-                print(f"{datetime.datetime.now()}: Temporary files deleted.")
+        if cache.get(event_id) is None:
+            print(f"{datetime.datetime.now()}: Event not processed. Starting event processing.")
+            
+            # handle emoji reactions
+            if reaction_type != "label":
+                print(f"{datetime.datetime.now()}: Reaction was of type :{reaction_type}:. Ignoring.")
             else:
-                print(f"{datetime.datetime.now()}: No file found.")
+                
+                conversation = client.conversations_replies(
+                    channel=channel_id,
+                    ts=timestamp
+                )
+                messages = conversation.get("messages")
+                root_message = messages[0]
+
+                if "files" in root_message:
+                    print(f"{datetime.datetime.now()}: File found.")
+                    file_id = root_message["files"][0]["id"]
+                    file_vtt = get_file_info(file_id)
+                    save_location = event_id + '.vtt'
+                    vtt_file_for_conversion = download_vtt_file(file_vtt, save_location)
+                    txt_file_output = event_id + ".txt"
+                    finished_txt_file = convert_vtt_to_labels(vtt_file_for_conversion, txt_file_output)
+                    client.files_upload_v2(
+                    channel=channel_id,
+                    thread_ts=timestamp,
+                    initial_comment="Thanks for using LabelMaker! Here's your labels file:",
+                    file=finished_txt_file
+                )
+                    os.remove(vtt_file_for_conversion)
+                    os.remove(finished_txt_file)
+                    print(f"{datetime.datetime.now()}: Temporary files deleted.")
+                else:
+                    print(f"{datetime.datetime.now()}: No file found.")
+            cache.set(event_id, True)
+            print(f"{datetime.datetime.now()}: Event processed.")
+
+        else:
+            print(f"{datetime.datetime.now()}: Event has already been processed.")
 
     return "OK"
 
 def get_file_info(file_id):
     print(f"{datetime.datetime.now()}: Attempting to get VTT file from Slack.")
     response = client.files_info(file=file_id)
-    while "vtt" not in response["file"]:
-        print(f"{datetime.datetime.now()}: Slack has not generated a VTT file yet. Trying again...")
-        response = client.files_info(file=file_id)
-    print(f"{datetime.datetime.now()}: Transcript generated, link found.")
-    vtt_link = response["file"]["vtt"]
+    if "vtt" not in response["file"]:
+        print(f"{datetime.datetime.now()}: Slack has not generated a VTT file yet. Trying again in 10 seconds.")
+        time.sleep(10.0)
+        get_file_info(file_id)
+    else: 
+        print(f"{datetime.datetime.now()}: Transcript generated, link found.")
+        vtt_link = response["file"]["vtt"]
     return vtt_link
 
 def download_vtt_file(url, save_path):
